connect to postman

+++ src/main/java/com/coffeeshop/server/model/MenuProduk.java
@@ -9,13 +9,13 @@ import lombok.Data;
 public class MenuProduk {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Long id; // Menggunakan Long untuk ID adalah standar yang bagus di Spring Boot
 
     @Column(name = "nama_produk", nullable = false)
     private String namaProduk;
 
     @Column(nullable = false)
-    private Double harga;
+    private Integer harga; // <-- Tipe data diubah menjadi Integer agar sinkron dengan INT di MySQL
 
     @Column(columnDefinition = "TEXT")
     private String deskripsi;
